test: provide ability to test js code
