+++ package.json
@@ -7,7 +7,10 @@
   "scripts": {
     "intro": "cat package.json | jq .scripts",
     "sass": "node-sass --output-style expanded promptly/static/scss/ -o promptly/static/css/",
-    "doc:img": "drawio docs/database.drawio -o docs/_static/database.png"
+    "doc:img": "drawio docs/database.drawio -o docs/_static/database.png",
+    "build": "NODE_ENV=production webpack",
+    "build:dev": "NODE_ENV=development webpack",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -34,7 +37,14 @@
   },
   "homepage": "https://promptly.readthedocs.io/en/latest/",
   "devDependencies": {
+    "@babel/core": "^7.23.2",
+    "@babel/preset-env": "^7.23.2",
+    "babel-loader": "^9.1.3",
     "draw.io-export": "^0.3.0",
-    "node-sass": "^9.0.0"
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "node-sass": "^9.0.0",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
   }
 }
