+++ promptly/templates/layout/base.html
@@ -47,7 +47,8 @@
     </div>
 
     <script src="{{ url_for('static', filename='js/app.js') }}"></script>
-    <script src="{{ url_for('static', filename='js/script.js') }}"></script>
+    <!-- TODO: production: script.min.js, development: script.js  -->
+    <script src="{{ url_for('static', filename='js/script.min.js') }}"></script>
     {% block body_js %}{% endblock %}
   </body>
 </html>
